[RSDK-10493] Update sdk version to something recent (#10)
I went with 0.70.0, released last week. This week's release has already needed a patch (and may or may not need more patches).

Everything compiles and the tests pass, but I am unsure how to actually try running it outside of production. but I kinda imagine that if it compiles, it probably still works.

The `Camera` interface lost its streaming capability, so I changed to use `.Image()` to get frames instead. 


* bump RDK version to 0.70.0 (from April 8)

* go mod tidy

* switch from streams to getting individual frames from the cameras, doesn't compile yet

* get it to compile again

* pr feedback

* add dependencies in the makefile so it rebuilds when something changes